requirements.txt
PyQt6>=6.4.0
PyMuPDF>=1.21.0
Pillow>=9.3.0
python-dotenv>=0.21.0
pywin32>=305; platform_system == "Windows"
requests>=2.28.1  # For file upload functionality
pyinstaller>=5.13.0  # For creating executable